Move jobs/ and tests/ to project root
